+++ packages/vsce/CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to the "cics-extension-for-zowe" extension will be documente
 
 ## Recent Changes
 
+- Enhancement: Added Recent Resources QuickPick for Inspect and Compare Resource commands. Shows last 5 resources per type with free-text entry support. [#649](https://github.com/zowe/cics-for-zowe-client/issues/649)
  - Enhancement: Adds hyperlinks to appropriate attributes in RI highlights. [#621](https://github.com/zowe/cics-for-zowe-client/issues/621)
  - BugFix: Render USS paths with a trailing slash (e.g. `JAVA_HOME=/usr/lpp/java/J8.0_64/`) as hyperlinks in the Resource Inspector. [#638](https://github.com/zowe/cics-for-zowe-client/issues/638)
  - Enhancement: Use SDK-provided methods to close and open local files. [#241](https://github.com/zowe/cics-for-zowe-client/issues/241)